+++ src/adapter/setup.ts
@@ -29,12 +29,11 @@ function getNextFootnoteId(): number {
   return 1 + parseInt(lastFootnoteId, 10)
 }
 
-function findFootnoteLinks({
-  anchorPattern,
-  scope = ''
-}: Settings): readonly HTMLAnchorElement[] {
+function findFootnoteLinks(
+  anchorPattern: RegExp
+): readonly HTMLAnchorElement[] {
   return Array.from(
-    document.querySelectorAll<HTMLAnchorElement>(scope + ' a[href^="#"]')
+    document.querySelectorAll<HTMLAnchorElement>('a[href^="#"]')
   ).filter(link => (link.href + link.rel).match(anchorPattern))
 }
 
@@ -148,7 +147,7 @@ export function createDocumentFootnotes(settings: Settings): RawFootnote[] {
   const { buttonTemplate, contentTemplate, numberResetSelector } = settings
   const offset = getNextFootnoteId()
 
-  return findFootnoteLinks(settings)
+  return findFootnoteLinks(settings.anchorPattern)
     .map(findRefBody(settings))
     .filter(isDefined)
     .map(hideOriginalFootnote)