+++ src/types.ts
@@ -13,7 +13,6 @@ export type Settings = Readonly<{
   footnoteSelector: string
   hoverDelay: number
   numberResetSelector?: string
-  scope?: string
 }>
 
 export type Footnote = Readonly<{